requirements.txt
Flask==2.3.3
Werkzeug==2.3.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
#psycopg2-binary==2.9.6  # Only needed for PostgreSQL
email-validator==2.0.0  # For email validation in forms
